Time: 0 ms (100%), Space: 20 MB (17.09%) - LeetHub

+++ 0229-majority-element-ii/0229-majority-element-ii.cpp
@@ -4,7 +4,7 @@ public:
         int n = nums.size();
         map<int,int>mpp;
         vector<int>ls;
-        int mn = int(n/3) + 1;
+        int mn = (n/3) + 1;
         for(int i = 0; i<n; ++i)
         {
             mpp[nums[i]]++;
